% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEstimateProportion.R
\name{apply_gate}
\alias{apply_gate}
\title{Apply gate_flowclust_1d gate}
\usage{
apply_gate(
  ff,
  channel,
  cutpoint_min,
  cutpoint_max,
  return_plot = TRUE,
  join = T,
  traditional = F
)
}
\arguments{
\item{ff}{A flowFrame object to be gated.}

\item{channel}{A character vector specifying the channel(s) to be used during gating.}

\item{cutpoint_min}{A numeric value that sets a minimum threshold for the cutpoint.
If a value is provided, any cutpoint below this value will be set to the given minimum value.
If NULL (default), there is no minimum cutpoint value.}

\item{cutpoint_max}{A numeric value that sets a maximum threshold for the cutpoint.
If a value is provided, any cutpoint above this value will be set to the given maximum value.
If NULL (default), there is no maximum cutpoint value.}

\item{return_plot}{A logical value indicating whether the result should include the plot with the gates.}

\item{join}{A logical value indicating whether the result should be improved by determining a target population through the given channels.}

\item{traditional}{A logical value indicating whether to use a traditional gating approach.}
}
\value{
A list containing the following components:
\item{plot}{A ggplot object showing the plot with applied gates (if return_plot is TRUE).}
\item{idt}{A logical vector indicating gated events.}
\item{channel}{The channels used for gating.}
\item{gate_bg}{A list of gates applied to each channel.}
}
\description{
Apply gate_flowclust_1d gate to one or two channels. If enough information about the population is kept and the join parameter is active,
a target gate_flowclust_2d is applied to increase resolution.
}
\examples{
apply_gate(ff = your_flowFrame_object, channel = c("FL1-A", "FL2-A"))

}
\keyword{flowCore}
