% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{cytokine_cutpoint}
\alias{cytokine_cutpoint}
\title{Calculate Cytokine Cutpoint from OpenCyto}
\usage{
cytokine_cutpoint(
  x,
  num_peaks = 1,
  ref_peak = 1,
  method = c("first_deriv", "second_deriv"),
  tol = 0.01,
  adjust = 1,
  side = "right",
  strict = TRUE,
  plot = FALSE,
  auto_tol = FALSE,
  ...
)
}
\arguments{
\item{x}{A numeric vector containing data values.}

\item{num_peaks}{Number of peaks to consider for cutpoint calculation.}

\item{ref_peak}{Reference peak for cutpoint calculation.}

\item{method}{Method for calculating cutpoint ("first_deriv" or "second_deriv").}

\item{tol}{Tolerance for calculating the cutoff point.}

\item{adjust}{Adjustment factor for peak detection.}

\item{side}{Side for cutpoint calculation ("left" or "right").}

\item{strict}{Logical indicating whether to strictly follow the cutoff point or not.}

\item{plot}{Logical indicating whether to plot the peak detection.}

\item{auto_tol}{Logical indicating whether to automatically set tolerance based on peak values.}

\item{...}{parameter to pass to deriv_density}
}
\value{
The calculated cutpoint.
}
\description{
This function calculates the cutpoint for a cytokine gate based on specified parameters. It is based on \code{.cytokine_cutpoint} from OpenCyto pacakge.
}
\details{
Finak, Greg, Frelinger, Jacob, Jiang, Wenxin, Newell, Evan W., Ramey, John, Davis, Mark M., Kalams, Spyros A., De Rosa, Stephen C., Gottardo, Raphael (2014). “OpenCyto: An Open Source Infrastructure for Scalable, Robust, Reproducible, and Automated, End-to-End Flow Cytometry Data Analysis.” PLoS Computational Biology, 10(8), e1003806.
}
