% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{deriv_density}
\alias{deriv_density}
\title{Calculate Derivative of Density from OpenCyto}
\usage{
deriv_density(
  x,
  deriv = 1,
  bandwidth = NULL,
  adjust = 1,
  num_points = 10000,
  ...
)
}
\arguments{
\item{x}{A numeric vector containing data values.}

\item{deriv}{Order of the derivative.}

\item{bandwidth}{Bandwidth for kernel density estimation.}

\item{adjust}{Adjustment factor for bandwidth.}

\item{num_points}{Number of points for estimation.}
}
\value{
A list with x and y components representing the calculated derivative.
}
\description{
This function calculates the derivative of the density function.
}
\details{
Finak, Greg, Frelinger, Jacob, Jiang, Wenxin, Newell, Evan W., Ramey, John, Davis, Mark M., Kalams, Spyros A., De Rosa, Stephen C., Gottardo, Raphael (2014). “OpenCyto: An Open Source Infrastructure for Scalable, Robust, Reproducible, and Automated, End-to-End Flow Cytometry Data Analysis.” PLoS Computational Biology, 10(8), e1003806.
}
