% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEstimateProportion.R
\name{do_ggcyto}
\alias{do_ggcyto}
\title{Create ggcyto plot}
\usage{
do_ggcyto(
  fC,
  channels,
  gates = NULL,
  logicle_chnls = NULL,
  main = NULL,
  legend = FALSE
)
}
\arguments{
\item{fC}{flowCore object.}

\item{channels}{Vector of length 2 specifying the channels to be plotted.
If only one channel is provided, \code{SSC-A} will be used as the second channel.}

\item{gates}{List of gates to be included in the ggcyto plot.}

\item{logicle_chnls}{Vector of channel names to be transformed using logicle scale.}

\item{main}{Main title for the plot.}

\item{legend}{Logical, whether to show the legend in the plot.}
}
\value{
A ggplot object.
}
\description{
Create a ggcyto plot based on two given channels (or one and by default \code{SSC-A} will be used as the second one). The ggcyto
could include gates and logicle scale.
}
\examples{
do_ggcyto(fC = your_flowCore_object, channels = c("FL1-A", "FL2-A"))

}
\keyword{plot}
