% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDA.R
\name{doDA}
\alias{doDA}
\title{Run Differential Abundance Analysis}
\usage{
doDA(sce_file, output, response, response_label)
}
\arguments{
\item{sce_file}{Path to the SingleCellExperiment (SCE) object file.}

\item{output}{Directory where the output files will be saved.}

\item{response}{Column name in the metadata indicating the response variable.}

\item{response_label}{Labels for the response variable levels.}
}
\value{
A data frame containing results of statistical tests or NA if the response variable is missing.
}
\description{
This function performs differential analysis and generates various plots based on the provided single-cell experiment data.
}
