% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBackgating.R
\name{doDensityBackgating}
\alias{doDensityBackgating}
\title{Perform density-based gating on flow cytometry data}
\usage{
doDensityBackgating(
  ff,
  filename,
  output.dir,
  chnl = c("FSC-A", "SSC-A"),
  channel_bg,
  logicle_chnls = NULL,
  sd.max_it = 0.75,
  min.pct_it = 0.01,
  target.fsc = 50000,
  target.ssc = 17500,
  min.ff_subset = 250
)
}
\arguments{
\item{ff}{FlowFrame object containing the flow cytometry data.}

\item{filename}{Name of the input file.}

\item{output.dir}{Directory where output files will be saved.}

\item{chnl}{Vector of channel names for scatter plots (default: c("FSC-A", "SSC-A")).}

\item{channel_bg}{Vector of channel names for background gating.}

\item{logicle_chnls}{Vector of channel names for logicle transformation (if applicable).}

\item{sd.max_it}{Standard deviation factor for adjusting gating percentages.}

\item{min.pct_it}{Minimum gating percentage for selection.}

\item{target.fsc}{Target FSC value for gating adjustment.}

\item{target.ssc}{Target SSC value for gating adjustment.}

\item{min.ff_subset}{Minimum number of events to keep in the subset.}
}
\value{
Number of selected events after gating.
}
\description{
This function performs density-based gating on flow cytometry data using various parameters.
}
\examples{
# Load required packages
library(ggcyto)
library(flowCore)

# Load flow cytometry data (replace with actual data)
data_file <- system.file("extdata", "sample_data.fcs", package="flowCore")
ff <- read.FCS(data_file)

# Define channel names and background gating channels
channel_bg <- c("CD3-FITC", "CD4-PE", "CD8-APC")

# Perform density-based gating
doDensityBackgating(ff, "sample_data.fcs", output.dir = "output_directory", channel_bg = channel_bg)

}
\seealso{
\code{\link{ggcyto::autoplot}}
\code{\link{ggplot2::ggsave}}
\code{\link{flowCore::filter}}
\code{\link{flowCore::write.FCS}}
\code{\link{openCyto::gate_tail_custom}}
\code{\link{openCyto::gate_flowclust_2d}}
}
\keyword{cytometry}
\keyword{density-based}
\keyword{flow}
\keyword{gating}
