% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEstimateProportion.R
\name{doEstimateProportion}
\alias{doEstimateProportion}
\title{Estimate Proportions and Create Plots}
\usage{
doEstimateProportion(
  filename,
  id = NULL,
  output.dir,
  info_panel,
  cutpoint_min = 1,
  cutpoint_max = 5e+05
)
}
\arguments{
\item{filename}{The path to the FCS file.}

\item{id}{A character identifier for the file (default: NULL, uses filename).}

\item{output.dir}{The output directory for plots and summary files.}

\item{info_panel}{A data frame containing information about patterns and labels.}

\item{cutpoint_min}{A numeric value setting a minimum threshold for gating.}

\item{cutpoint_max}{A numeric value setting a maximum threshold for gating.}
}
\description{
This function reads FCS files, applies gating based on provided patterns, and estimates proportions for each pattern.
It also generates plots and outputs a summary file.
}
\examples{
doEstimateProportion(
  filename = "path/to/your/fcs_file.fcs",
  output.dir = "output_directory",
  info_panel = info_panel_data_frame
)

}
\keyword{flowCore}
