% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPreprocessing.R
\name{doPreprocessing}
\alias{doPreprocessing}
\title{Perform Preprocessing on Flow Cytometry Data}
\usage{
doPreprocessing(file, filename, output, report = TRUE)
}
\arguments{
\item{file}{Path to the input FCS file.}

\item{filename}{Desired name for the output FCS file (without extension).}

\item{output}{Directory where output files and reports will be saved.}

\item{report}{Logical indicating whether to generate a preprocessing report. Default is TRUE.}
}
\value{
If successful, the function performs preprocessing steps and generates output files.
}
\description{
This function performs a series of preprocessing steps on flow cytometry data, including quality control, compensation,
singlet gating, and generating a preprocessing report.
}
\examples{
\dontrun{
doPreprocessing(file = "path/to/input.fcs", filename =
  "preprocessed_input.fcs", output = "output_directory")
}

}
