% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSEDA.R
\name{doSEDA}
\alias{doSEDA}
\title{Extracts and processes FCS files with optional metadata and marker name.}
\usage{
doSEDA(fcs_path, metadata = NULL, maker_name = NULL)
}
\arguments{
\item{fcs_path}{A character vector of file paths to FCS files.}

\item{metadata}{Optional data frame containing metadata for FCS files.}

\item{maker_name}{Optional name for the marker to be added in metadata.}
}
\value{
A processed single-cell experiment (SCE) object.
}
\description{
This function extracts data from FCS files, processes it, and returns a processed
single-cell experiment (SCE) object. It supports filtering and transformation of
the data, with the option to provide metadata and a marker name for the experiment.
}
\examples{
doSEDA(fcs_path = c("file1.fcs", "file2.fcs"), metadata = meta_data, maker_name = "Marker1")


}
