% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSEDA.R
\name{downsampling_with_outliers}
\alias{downsampling_with_outliers}
\title{Downsampling with Proportional Outliers}
\usage{
downsampling_with_outliers(expression_matrix, target_cells)
}
\arguments{
\item{expression_matrix}{A matrix of gene expression data where rows represent cells and columns represent genes.}

\item{target_cells}{The desired number of cells after downsampling.}
}
\value{
A downsampled gene expression matrix.
}
\description{
This function performs downsampling on a single-cell gene expression matrix while maintaining the proportion of outliers.
}
\examples{
# Suppose 'expression_matrix' is your gene expression matrix and 'target_cells' is the desired number of cells after downsampling
sampled_expression_matrix <- downsampling_with_outliers(expression_matrix, target_cells)


}
