% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPreprocessing.R
\name{filterSinglets}
\alias{filterSinglets}
\title{Filter Singlet Cells from Flow Cytometry Data}
\usage{
filterSinglets(fC, chnl = c("FSC-A", "FSC-H"))
}
\arguments{
\item{fC}{A flow cytometry dataset (typically a flowFrame object) containing the raw data from flowCore package.}

\item{chnl}{A character vector of length 2 specifying the channels to use for singlet gating.
The default is c("FSC-A", "FSC-H").}
}
\value{
A flow cytometry dataset with only singlet cells based on the gating results.
}
\description{
This function filters singlet cells from flow cytometry data using the PeacoQC package
to remove doublets and the openCyto package for singlet gating. It is a double checking for singlets by looking \code{FSC-A} and \code{FSC-H} channels.
}
\examples{
\dontrun{
filtered_data <- filterSinglets(fC = my_flow_data, chnl = c("FSC-A", "FSC-H"))
}

}
