% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{gate_tail_custom}
\alias{gate_tail_custom}
\title{Custom Gate with Tail from OpenCyto}
\usage{
gate_tail_custom(
  fr,
  channel,
  filterId = "",
  num_peaks = 1,
  ref_peak = 1,
  strict = TRUE,
  tol = 0.01,
  side = "right",
  min = NULL,
  max = NULL,
  bias = 0,
  positive = TRUE,
  ...
)
}
\arguments{
\item{fr}{A flowFrame containing flow cytometry data.}

\item{channel}{Name of the channel to create the gate on.}

\item{filterId}{ID for the created gate.}

\item{num_peaks}{Number of peaks to consider for tail gate calculation.}

\item{ref_peak}{Reference peak for tail gate calculation.}

\item{strict}{Logical indicating whether to strictly follow the cutoff point or not.}

\item{tol}{Tolerance for calculating the cutoff point.}

\item{side}{Side for gate creation ("left" or "right").}

\item{min}{Minimum value for truncating data before gate creation.}

\item{max}{Maximum value for truncating data before gate creation.}

\item{bias}{Bias to adjust the cutoff point.}

\item{positive}{Logical indicating whether gate should be positive or not.}
}
\value{
A rectangleGate representing the custom gate.
}
\description{
This function creates a custom gate with a tail based on specified parameters.
}
\details{
Finak, Greg, Frelinger, Jacob, Jiang, Wenxin, Newell, Evan W., Ramey, John, Davis, Mark M., Kalams, Spyros A., De Rosa, Stephen C., Gottardo, Raphael (2014). “OpenCyto: An Open Source Infrastructure for Scalable, Robust, Reproducible, and Automated, End-to-End Flow Cytometry Data Analysis.” PLoS Computational Biology, 10(8), e1003806.
}
