% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBackgating.R
\name{get_contour}
\alias{get_contour}
\title{Get Contour}
\usage{
get_contour(ff)
}
\arguments{
\item{ff}{An object of class 'flowFrame' representing the flow cytometry data.}
}
\value{
A list of contour lines containing x and y coordinates.
}
\description{
Generate and plot contour lines of the density estimate for given FSC-A and SSC-A channels.
}
\details{
This function calculates the kernel density estimate of the given FSC-A and SSC-A channels
and generates contour lines from the density estimate. It iteratively refines the contour lines
to identify dense regions and plots them using ggplot2.
}
\examples{
# Load the required libraries
library(flowCore)
library(ggplot2)
library(ggplotify)

# Load an FCS file using flowCore
fcs_file <- read.FCS("path/to/your/fcsfile.fcs")

# Generate and plot contour lines
contour_lines <- get_contour(fcs_file)

}
