% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSEDA.R
\name{min_max_sampling}
\alias{min_max_sampling}
\title{Perform Min-Max Gene Sampling on Expression Matrix}
\usage{
min_max_sampling(expression_matrix, target_cells)
}
\arguments{
\item{expression_matrix}{A numeric matrix where rows represent genes and columns represent cells.}

\item{target_cells}{The desired number of cells to be selected from the expression matrix.}
}
\value{
A vector of indices representing the selected cells.
}
\description{
This function performs Min-Max gene sampling on an expression matrix to select a subset of cells
that represent a broad expression range across genes. The algorithm aims to maximize the diversity
of gene expression profiles among the selected cells.
}
\examples{
expression_matrix <- matrix(runif(1000), nrow = 100, ncol = 10)
selected_cells <- min_max_sampling(expression_matrix, target_cells = 5)

}
