% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seacellsR.R
\name{run_min_max_sampling}
\alias{run_min_max_sampling}
\title{Run MinMax-Sampling using SEACells}
\usage{
run_min_max_sampling(
  file_counts,
  output,
  n_SEACells,
  n_waypoint_eigs = 5,
  n_comps_pca = 10,
  build_kernel_on = "X_umap"
)
}
\arguments{
\item{file_counts}{File with raw counts}

\item{n_SEACells}{Number of cells to keep}

\item{n_components}{Number of components}

\item{...}{Other arguments passed to the Python function}
}
\value{
Path to the file with the selected cell
}
\description{
Run MinMax-Sampling using SEACells
}
