% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_function.R
\name{run_Preprocessing}
\alias{run_Preprocessing}
\title{Preproccessing Function}
\usage{
run_Preprocessing(file, filename, output)
}
\arguments{
\item{file}{path to the file FCS to run.}

\item{filename}{filename for the cleaned FCS.}

\item{output}{path to the location for the cleaned FCS.}
}
\description{
This function allows run preprocessing analysis for raw FCS file. It does the compensation and
the QC (remove doublets and anomalies)
}
\examples{
run_Preprocessing()
}
\keyword{preprocessing}
