% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDA.R
\name{runDA}
\alias{runDA}
\title{Run Differential Analysis from Preprocessed Data}
\usage{
runDA(output, response, response_label)
}
\arguments{
\item{output}{Directory where the output files will be saved.}

\item{response}{Column name in the metadata indicating the response variable.}

\item{response_label}{Labels for the response variable levels.}
}
\value{
A data frame containing results of statistical tests or NA if the response variable is missing.
}
\description{
This function runs the differential analysis process using a preprocessed SingleCellExperiment (SCE) object and generates various plots based on the provided information.
}
