% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPreprocessing.R
\name{runPreprocessing}
\alias{runPreprocessing}
\title{Run Preprocessing Function}
\usage{
runPreprocessing(
  metadata,
  output,
  report = TRUE,
  cluster = NULL,
  log_file = "Log_file_error_raw2clean.txt"
)
}
\arguments{
\item{metadata}{A data frame with a column named "filename" containing paths to the raw FCS files.}

\item{output}{Path to the location where the cleaned FCS files will be saved.}

\item{report}{Logical indicating whether to report the number of anomalies and doublets. Default is TRUE.}

\item{cluster}{Number of parallel workers to be used for processing. Default is NULL.}

\item{log_file}{Path to the log file where error messages will be written.}
}
\value{
This function runs the preprocessing analysis and generates cleaned FCS files and reports.
}
\description{
This function runs the preprocessing analysis for raw FCS files. It performs compensation and quality control,
including removing doublets and anomalies.
}
