% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPreprocessing.R
\name{simplify_flowCore}
\alias{simplify_flowCore}
\title{Simplify flowCore Object}
\usage{
simplify_flowCore(filename, keep = NULL)
}
\arguments{
\item{filename}{The path to the FCS file to be simplified.}

\item{keep}{A character vector specifying the names of channels to keep. Default is NULL.}
}
\value{
If \code{keep} is NULL, returns the path to the simplified FCS file. If \code{keep} is specified,
returns the path to the simplified FCS file after keeping only specified channels.
}
\description{
This function simplifies a flowCore object by removing empty channels and renaming
channels based on their descriptions. It also allows for keeping specific channels if needed.
}
\examples{
\dontrun{
simplified_file <- simplify_flowCore(filename = "path/to/input.fcs")
simplified_kept_file <- simplify_flowCore(filename = "path/to/input.fcs", keep = c("FSC-A", "SSC-A"))
}

}
