% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seacellsR.R
\name{run_min_max_sampling}
\alias{run_min_max_sampling}
\title{Run MinMax-Sampling using SEACells}
\usage{
run_min_max_sampling(
  file_counts,
  output,
  n_SEACells,
  n_waypoint_eigs = 5,
  n_comps_pca = 10,
  build_kernel_on = "X_umap",
  conda_env = NULL
)
}
\arguments{
\item{file_counts}{Path to the file containing counts data.}

\item{output}{Directory where the output files will be saved.}

\item{n_SEACells}{Number of SEACells.}

\item{n_waypoint_eigs}{Number of waypoint eigenvectors (default is 5).}

\item{n_comps_pca}{Number of PCA components (default is 10).}

\item{build_kernel_on}{Build kernel on which data (default is "X_umap").}

\item{conda_env}{Conda environment to use for running SEACells.}
}
\value{
Path to the file containing the indices used for min-max sampling.
}
\description{
This function performs min-max sampling using SEACells or falls back to a basic sampling method if there's an error with the conda environment.
}
\examples{
run_min_max_sampling(
  file_counts = "counts.txt",
  output = "output_folder",
  n_SEACells = 100,
  n_waypoint_eigs = 5,
  n_comps_pca = 10,
  build_kernel_on = "X_umap",
  conda_env = "my_env"
)

}
