% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSEDA.R
\name{runSEDA}
\alias{runSEDA}
\title{Run SEDA Analysis}
\usage{
runSEDA(
  output,
  metadata = NULL,
  marker = FALSE,
  downsampling = "random",
  k_downsampling = 1e+05,
  seed = NULL,
  batch = FALSE
)
}
\arguments{
\item{output}{Output directory for saving the analysis results.}

\item{metadata}{Optional data frame containing additional metadata information.}

\item{marker}{Whether to perform marker-specific analysis.}

\item{downsampling}{The type of downsampling to perform ("random", "minMaxSamplingSEACELLS", "minMaxSampling", or none).}

\item{k_downsampling}{Percentage of umber of cells to downsample for each ID.}

\item{seed}{Seed for reproducibility if downsampling is used.}

\item{batch}{Whether to perform batch effect correction.}
}
\description{
This function performs the SEDA analysis on FCS data, including optional marker-specific analysis and downsampling.
}
